requirements.txt
streamlit>=1.28.0
openai>=1.0.0
python-dotenv==1.0.0
audio-recorder-streamlit==0.0.10
requests>=2.31.0
pandas>=1.5.0,<3.0.0
numpy>=1.24.0,<2.0.0
openpyxl>=3.1.0
spacy>=3.7.0,<3.8.0
spacy-lookups-data==1.0.5
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_sm-0.5.4.tar.gz
langchain==0.3.27
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-chroma>=0.1.2
chromadb>=0.5.0,<0.6.0
huggingface_hub>=0.19.0
transformers>=4.30.0,<4.40.0
torch>=2.0.0,<2.5.0
